pyproject.toml
[project]
name = "rlm"
version = "0.1.0"
description = "Python implementation of Recursive Language Models - process arbitrarily long contexts via REPL + recursive LLM calls"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Repository = "https://github.com/Ray0907/rlm"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rlm"]
